+++ server/client/room.go
@@ -1,4 +1,4 @@
-package cli
+package client
 
 import (
 	"time"
